Fix compatibility with MSYS2/MinGW-w64.
The 'SDL2_main' library is such a damn pain, and now it's gotten even
more annoying because you have to include a bloody header in the file
that declares the 'main' function. This would be a lot easier if SDL2
defined a preprocessor macro so that code can just tell when to
include the header without me needing to do it manually.