Restore DOS SFX behaviour.
Now SFX are not all permanently cached on startup. This is done by
restoring dummied-out code from the DOS version of the engine.

I've also changed the code so that the sample count is read from the
SFX's header rather than being derived from the size of the SFX's
lump.