Add support for 'full-colour' rendering.
The screenshot code needs fixing.

Normally Doom is limited to 256 colours, and uses proxy palettes to
translate sprites from brighter colours to darker colours. This is
what gives Doom is poor colour-depth.

However, these proxy palettes can be made to harmlessly pass-through
to an expanded palette, which surpasses the 256 colour limit. That is
what this commit does. With this, the game renders without any colour
loss.

Instead of using an array of precalcualted palettes and a series of
translation tables ("colour maps"), the full-colour renderer only
uses the main palette, and darkens and tints it as needed, allowing
it to create the required colour rather than alias them to existing
colours.

In effect, the framebuffer has gone from 8-bit to 16-bit.