Greatly-optimise fixed-point division.
Just today I realised that the console ports of Doom would have had
to do the same overhaul to the fixed-point arithmetic that I did, as
they could not rely on the x86's 64-bit division instruction.

To my surprise, the Atari Jaguar port of Doom appears to use the same
general algorithms as my code does. However, the division function is
substationally different, making use of some clever optimisations.

According to Visual Studio's profiler, this takes `M_FixedDiv` from
6.46% CPU time to 0.46%. It is one hell of a performance improvement,
and it produces identical results to a proper 64-bit division.